* kiss_chars.h   - Forgot to add this to repo when splitting it out * dvopus-tx.c   - Made the frame iterator a 16-bit value for breathing room * dvopus-rx.c   - Initial checkin   - Does not decode audio yet. Wanted to get working KISS decoding     checked in before continuing. * Makefile   - Add support for building receiver

+++ dvopus-tx.c
@@ -26,7 +26,7 @@ int oerr;                       // Error code from the encoder
 opus_int32 opus_len;            // Return length from the encoder
 uint8_t out_buf[256];           // Radio frame buffer
 uint32_t csum;                  // Checksum of output frame
-uint8_t i;                      // generic iterator
+uint16_t i;                     // generic iterator
 uint16_t frame_seq;             // Frame sequence number
 
 const char magic[] = "DVOp";
